Added outsideApiCalls file to improve separation of concerns.
