+++ server_friendle/routes/signup.js
@@ -1,34 +1,8 @@
 const router = require("express").Router();
 const knex = require("knex")(require("../knex_db/knexfile").development);
 const bcrypt = require("bcryptjs");
-const axios = require("axios");
-const fs = require("fs");
 const validate = require("../utils/validators");
-
-// Generating a random seed for unique user icons
-const seedrandom = require("seedrandom");
-const generator = seedrandom();
-
-const userIconCall = (username) => {
-  const randomSeed = generator();
-  const filePath = `/assets/user_icons/${username}_icon.svg`;
-  axios
-    .get(`https://avatars.dicebear.com/api/identicon/${randomSeed}.svg`)
-    .then((res) => {
-      fs.writeFile(
-        `./assets/user_icons/${username}_icon.svg`,
-        res.data,
-        (err) => {
-          if (err) console.log(err);
-          console.log("It worked");
-        }
-      );
-    })
-    .catch((err) => {
-      console.log(err);
-    });
-  return filePath;
-};
+const outsideApi = require("../utils/outsideApiCalls");
 
 router.post("/", (req, res) => {
   const data = req.body;
@@ -36,7 +10,7 @@ router.post("/", (req, res) => {
   const formattedUsername = username.replace(/ /g, "_");
   if (validate.emailCheck(email) && username) {
     const hashedPassword = bcrypt.hashSync(password, 10);
-    const avatar = userIconCall(formattedUsername);
+    const avatar = outsideApi.userIconCall(formattedUsername);
     console.log(avatar);
     knex("users")
       .insert({ avatar, username, email, password: hashedPassword })
