+++ server_friendle/utils/validators.js
@@ -1,6 +1,4 @@
-const emailCheck = (email) => {
+exports.emailCheck = (email) => {
   const regex = new RegExp("[a-z0-9]+@[a-z]+.[a-z]{2,3}");
   return regex.test(email);
 };
-
-exports.emailCheck = emailCheck;
